+++ app/src/main/java/com/hampshirewolves/aeroatlas/ui/fragments/profilepage/ProfileFragment.java
@@ -1,22 +1,32 @@
 package com.hampshirewolves.aeroatlas.ui.fragments.profilepage;
 
+import android.content.Intent;
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.ImageView;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.hampshirewolves.aeroatlas.ui.mainactivity.MainActivity;
 import com.hampshirewolves.aeroatlas.R;
-import com.hampshirewolves.aeroatlas.ui.fragments.UserPageFragment;
-import com.hampshirewolves.aeroatlas.ui.fragments.discoverpage.DiscoverPageFragment;
+
 
 public class ProfileFragment extends Fragment {
 
-    private FirebaseAuth firebaseAuth;
+    TextView userPageEmailTextView;
+    Button userPageLogoutButton;
+
+    ImageView userPageBackButton;
+    TextView signupPageEmailValue;
 
     public ProfileFragment() {
         // Required empty public constructor
@@ -24,38 +34,41 @@ public class ProfileFragment extends Fragment {
 
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-
-        firebaseAuth = FirebaseAuth.getInstance();
-
-        checkUserStatus();
-
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_profile, container, false);
+        return inflater.inflate(R.layout.fragment_profile_page, container, false);
     }
 
-    private void checkUserStatus() {
-        FirebaseUser firebaseUser = firebaseAuth.getCurrentUser();
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        userPageEmailTextView = view.findViewById(R.id.userPageEmailTextView);
+        userPageLogoutButton = view.findViewById(R.id.userPageLogoutButton);
+        userPageBackButton = view.findViewById(R.id.userPageBackButton);
+        signupPageEmailValue = view.findViewById(R.id.signupPageEmailValue);
+
+        FirebaseUser firebaseUser = FirebaseAuth.getInstance().getCurrentUser();
 
         if (firebaseUser != null) {
-            navigateToUserPage();
+            userPageEmailTextView.setText(firebaseUser.getEmail());
+            signupPageEmailValue.setText(firebaseUser.getEmail());
         } else {
-            navigateToDiscoverPage();
+            userPageEmailTextView.setText("No email available");
         }
-    }
 
-    private void navigateToUserPage() {
-        getParentFragmentManager()
-                .beginTransaction()
-                .replace(R.id.baseFragment, new UserPageFragment())
-                .commit();
-    }
+        userPageLogoutButton.setOnClickListener(v -> {
+            FirebaseAuth.getInstance().signOut();
+            Toast.makeText(requireContext(), "Successfully logged out",
+                    Toast.LENGTH_SHORT).show();
+            Intent intent = new Intent(getActivity(), MainActivity.class);
+            startActivity(intent);
+        });
 
-    private void navigateToDiscoverPage() {
-        getParentFragmentManager()
-                .beginTransaction()
-                .replace(R.id.baseFragment, new DiscoverPageFragment())
-                .commit();
+        userPageBackButton.setOnClickListener(v -> {
+            Intent intent = new Intent(getActivity(), MainActivity.class);
+            startActivity(intent);
+        });
+
     }
 }
