+++ app/src/main/res/values/themes.xml
@@ -4,7 +4,6 @@
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
         <item name="android:navigationBarColor">@color/white</item>
-        <item name="android:statusBarColor">@color/white</item>
     </style>
 
     <style name="Theme.AeroAtlas" parent="Base.Theme.AeroAtlas" />
